+++ src/main/java/com/whereIsMyMoney/dao/BillDao.java
@@ -5,8 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BillDao extends JpaRepository<Bill, Integer> {
+public interface BillDao extends JpaRepository<Bill, Long> {
+
     List<Bill> findByShopName(String name);
+
     List<Bill> findByUserName(String name);
-    List<Bill> findByUserId(int id);
+
+    List<Bill> findByUserId(Long id);
 }
