+++ src/main/java/com/whereIsMyMoney/dao/ProductDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductDao extends JpaRepository<Product, Integer> {
+public interface ProductDao extends JpaRepository<Product, Long> {
     Product findByName(String name);
     List<Product> findByCategoryName(String name);
     List<Product> findByNameStartingWith(String name);
