+++ src/main/java/com/whereIsMyMoney/dao/PurchaseDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PurchaseDao extends JpaRepository<Purchase, Integer> {
-    Purchase findByProductId(int id);
-    List<Purchase> findByBillId(int id);
+public interface PurchaseDao extends JpaRepository<Purchase, Long> {
+    Purchase findByProductId(Long id);
+    List<Purchase> findByBillId(Long id);
 }
