+++ src/main/java/com/whereIsMyMoney/dao/ShopDao.java
@@ -3,6 +3,6 @@ package com.whereIsMyMoney.dao;
 import com.whereIsMyMoney.domain.Shop;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShopDao extends JpaRepository<Shop, Integer> {
+public interface ShopDao extends JpaRepository<Shop, Long> {
     Shop findByName(String name);
 }
